backend/requirements.txt
spacy
python-docx
pymupdf
transformers
sentence-transformers
keybert
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pandas